+++ database/migrations/2025_06_27_061845_create_purchase_items_table.php
@@ -11,21 +11,23 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('purchase_items', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('purchase_id')->constrained('purchases')->onDelete('cascade');
-            $table->foreignId('product_id')->constrained('products')->onDelete('restrict');
-            $table->unsignedInteger('qty');
-            $table->decimal('harga_beli', 15, 0);
-            $table->decimal('diskon', 15, 0)->default(0);
-            $table->foreignId('taxe_id')->nullable()->constrained('taxes');
-            $table->decimal('subtotal', 15, 0);
-            $table->timestamps();
+        if (!Schema::hasTable('purchase_items')) {
+            Schema::create('purchase_items', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('purchase_id')->constrained('purchases')->onDelete('cascade');
+                $table->foreignId('product_id')->constrained('products')->onDelete('restrict');
+                $table->unsignedInteger('qty');
+                $table->decimal('harga_beli', 15, 0);
+                $table->decimal('diskon', 15, 0)->default(0);
+                $table->foreignId('taxe_id')->nullable()->constrained('taxes');
+                $table->decimal('subtotal', 15, 0);
+                $table->timestamps();
 
-            $table->index('purchase_id');
-            $table->index('product_id');
-            $table->index('taxe_id');
-        });
+                $table->index('purchase_id');
+                $table->index('product_id');
+                $table->index('taxe_id');
+            });
+        }
     }
 
     /**
