+++ database/migrations/2025_09_18_113959_create_serial_numbers_table.php
@@ -11,18 +11,20 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('serial_numbers', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('product_id')->constrained('products')->onDelete('cascade');
-            $table->string('nomor_seri');
-            $table->enum('status', ['Tersedia', 'Terjual', 'Rusak', 'Hilang'])->default('Tersedia');
-            $table->foreignId('purchase_id')->nullable()->constrained('purchases');
-            $table->foreignId('item_sale_id')->nullable()->constrained('sale_items')->onDelete('set null');
-            $table->timestamps();
+        if (!Schema::hasTable('serial_numbers')) {
+            Schema::create('serial_numbers', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('product_id')->constrained('products')->onDelete('cascade');
+                $table->string('nomor_seri');
+                $table->enum('status', ['Tersedia', 'Terjual', 'Rusak', 'Hilang'])->default('Tersedia');
+                $table->foreignId('purchase_id')->nullable()->constrained('purchases');
+                $table->foreignId('item_sale_id')->nullable()->constrained('sale_items')->onDelete('set null');
+                $table->timestamps();
 
-            $table->unique(['product_id', 'nomor_seri']);
-            $table->index('status');
-        });
+                $table->unique(['product_id', 'nomor_seri']);
+                $table->index('status');
+            });
+        }
     }
 
     /**
