+++ database/migrations/2025_10_09_024746_create_product_promotion_table.php
@@ -11,12 +11,14 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('product_promotion', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('promotion_id')->constrained()->onDelete('cascade');
-            $table->foreignId('product_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('product_promotion')) {
+            Schema::create('product_promotion', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('promotion_id')->constrained()->onDelete('cascade');
+                $table->foreignId('product_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
